requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==41.0.7
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.0
django-allauth==0.60.1
django-jazzmin==2.6.0
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
oauthlib==3.2.2
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.8.0
urllib3==2.1.0
vonage==3.13.0
vonage-jwt==1.1.0
wrapt==1.16.0
yarl==1.9.4
